requirements.txt
fastapi
uvicorn
python-multipart
gunicorn
tensorflow
Pillow
numpy
requests